requirements.txt
-e git+http://github.com/mitsuhiko/flask.git#egg=Flask-dev
Flask-RESTful==0.3.1
Flask-SQLAlchemy==2.0
SQLAlchemy==0.9.8
psycopg2==2.6
tornado
python-daemon
